♻️  zb: Split AuthMechanism code into its own mod
